modify worke.js

+++ myquery/html5/Worker.js
@@ -2,50 +2,57 @@
 
 myQuery.define("html5/Worker", function ($, undefined) {
     "use strict"; //启用严格模式
-    if (window.Worker) {
-       
-        return {
-            worker: new window.Worker($.getPath("html5/_Worker"))
+
+    function MyWorker(path) {
+        if (window.Worker) {
+            this.worker = new window.Worker(path || $.getPath("html5/_Worker"));
+        }
+    }
+
+    MyWorker.prototype = {
+        constructor: MyWorker
             , addHandler: function (type, fun) {
                 /// <summary>添加worker事件</summary>
                 /// <param name="type" type="String">事件名</param>
                 /// <param name="fun" type="Function">方法</param>
                 /// <returns type="self" />
-                this.worker.addEventListener(type, fun, false);
+                this.worker && this.worker.addEventListener(type, fun, false);
                 return this;
             }
-
+            , on: function (type, fun) {
+                return this.addHandler(type, fun);
+            }
             , onError: function (fun) {
                 /// <summary>添加error事件</summary>
                 /// <param name="fun" type="Function">方法</param>
                 /// <returns type="self" />
-                return this.addHandler("error", fun);
+                return this.on("error", fun);
             }
             , onMessage: function (fun) {
                 /// <summary>添加获得数据事件</summary>
                 /// <param name="fun" type="Function">方法</param>
                 /// <returns type="self" />
-                return this.addHandler("message", fun);
+                return this.on("message", fun);
             }
 
-            , postMessage: function (todo, paras, context) {
+            , postMessage: function (todo, context, paras) {
                 /// <summary>发送计算函数</summary>
                 /// <param name="todo" type="Function">方法</param>
-                /// <param name="paras" type="Array">参数</param>
                 /// <param name="context" type="Object">作用域</param>
+                /// <param name="paras" type="Array">参数</param>
                 /// <returns type="self" />
-                this.worker.postMessage({
+                this.worker && this.worker.postMessage({
                     todo: todo.toString()
-                    , paras: paras || []
+                    , paras: $.argToArray(arguments, 2)
                     , context: context || window
                 });
                 return this;
             }
            , terminate: function () {
-               this.worker.terminate();
+               this.worker && this.worker.terminate();
                return this;
            }
-        };
-    }
-    return 0;
+       }
+
+    return MyWorker;
 });
